Changes made for project 3.
