+++ logs/debug.log
@@ -1,1 +1,1 @@
-11-01-2024 19:36:38.390 DEBUG [Main] - Hello World!
+11-15-2024 13:16:14.186 DEBUG [EscapeRoomGame] - Starting Up!
